+++ src/components/Admin/Charts/ChartUtils.ts
@@ -1,0 +1,7 @@
+import { Emotion } from "../../../models";
+
+export const emotionToNumeric = (emotion: Emotion) => ({
+    ":-)": 2,
+    ":-|": 1,
+    ":-(": 0,
+}[emotion])
